Extract remaining glTF extensions and accessor utilities from glb parser
